Finished "new" Call behavior and other changes
- Now, we're fully using the call's statuses instead of just
add/remove them from a list.
- A new status was added: finished. This was added to differentiate
when a call was just 'missed', or it was 'answered' and then, 'finished'.
- Some TODOs were added related with try/except. Maybe it will be dealt
in the next improvements (if time allows).
- New function created: get_call_on_queue(). Since we're not just
adding/removing the call from a list, we need to go through the list
and check the call's status in order to know whether it's on queue or not.

Signed-off-by: Rodrigo R. Galvao <rodrigorosattig@gmail.com>